added JPA

+++ new_version/src/main/java/pl/krzysiek/services/AccountService.java
@@ -1,6 +1,7 @@
 package pl.krzysiek.services;
 
 import pl.krzysiek.domain.Account;
+
 import java.util.List;
 
 
@@ -8,7 +9,7 @@ public interface AccountService {
 
     public List<Account> findAll();
     public void addAccount(Account acc);
-    public Account getById(long id);
+    public Account getById(int id);
     public Integer checkNickExists(String name);
     public int checkPassword(String email, String password);
     public Integer checkEmailExists(String email);
